requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic-settings==2.6.0
alembic==1.14.0
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.9.0
pytest==8.2.2
httpx==0.27.0
